requirements.txt
mysql-connector-python
Flask
bcrypt
flask-cors
cryptography
PyJWT
cuid2